Making Facenet faster by preloading model and exposing via RPC for Node.js
Example code for an RPC server that preloads Facenet with a desired
model and shows how to call that in Node.js for real world usage.
Reduces time to classification from ~11 seconds to <100 ms on a regular
Google Compute Engine CPU instance which is a big performance boost for
people looking to deploy this awesome work at scale.